Modify complete hook to sign certificates for RBAC
See issue #365

Previously, if RBAC was enabled on the Krake API, an Application
deployed with the "complete" hook could not be stopped using this hook.
Indeed, only the CA certificate was given to the Application. Only
having this CA did not allow the Application requests to the complete
hook endpoint to be authenticated. Thus, the request was considered
being sent anonymously. This is of course rejected by the RBAC
mechanism.

To allow the Applications to authenticate, a new certificate is
generated for each new Application. This process is done in the mangling
by the "complete" hook handler. The new certificate holds a specific CN,
which has been set in the configuration. If RBAC is enabled, this CN can
act as a user which has the rights to send requests to the "complete"
hook endpoints. These rights have to be given by adding a Role and a
RoleBinding to the database. The user in the RoleBinding needs to match
the one given in the configuration.

It means however that a certificate dedicated to signing new ones have
to be provided by the Krake administrator. Also, it means that in the
Application, the "complete" hook mangling adds the CA and the signing
certificate as CA bundle, and the generated certificate and its key. All
three are needed for the Application to be authenticated and authorized
by the Krake API.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>