Add role and its binding for the complete hook
See issue #365

Adds a Role and its RoleBinding to the elements to bootstrap. This
binding adds a role for the default "complete" hook user, which is the
user set in the certificate generated for a mangled Application.

This role allows the certificate user to delete Applications through the
"complete" hook endpoint.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>