Change the Magnum template field to immutable
On the MagnumCluster status, the template field which should be
immutable, and thus only set once by the API (during a binding in this
case), was wrongly not set.

To prevent changes in the Krake MagnumCluster resources's template field
which would not have been propagated, the field is set to immutable, as
it should have been.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>