Remove possibility to update without change
See issue #404

Previously, a user could send an update request for a resource to the
API, with the exact same body as what was currently in the database.
This would simply update the `metadata.modified` timestamp of the
resource.

To add some validation, this part has been changed to reject any update
that does not modify the ressource. This could prevent potential errors
of administrator trying to send update requests with wrong bodies.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>