Add resource and operation filtering in Generator
See issue #303

Many resources and operations are present in the templates for the API,
client and test generation. To allow for more flexibility and to also
ease development, the operations or resource to be displayed can now be
chosen. This prevents printing out a whole large file if only some parts
need debugging or be changed by the user.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>