Prevent empty factory list from failing black
See issue #303

When generating tests, factories are sometimes used. These factories
exist mostly for the Krake persistent resources, but not for the
resource lists for instance. Thus they do not need to be imported.
However, as the filtering was done on the template directly, the
"import" statement was present, but if no factory classes had to be
imported, the statement was empty and black returned an error.

Thus, the filtering is now done on the definition side, to be able to
compute beforehand if any factory class need to be imported.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>