Move the API generator and def outside of Krake
See issue #303

Previously, the API code generator was present in the Krake directory,
but was not used by Krake. Almost the same for the API definitions,
which are only used at some specific spots.

Because the generation of the code is not completely a part of Krake,
the generator is moved to its own directory on the repository's root.
However, no other modification of the rest of the code to adapt to the
shift has been performed.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>