Restrict the context to one in rok's cluster CRUD
Previously, when creating clusters with rok, several kubeconfig contexts
could be specified. For each one of these contexts present in the
kubeconfig file, a separate Cluster resource was created. However, when
updating the resource, no context could be specified, and the user had
to process the kubeconfig file him- or herself to update the cluster.
Also, if the certificates where not present in the kubeconfig file as
base64 encoding of their content, but as paths, the certificates path
were not processed, and the kubeconfig file became useless, because the
path were not necessarily the same in the Krake deployment.

This has been simplified, by only allowing one context in the creation
and update of the Cluster resources with the kubeconfig files. In those
cases, the files are processed in the same way, which ensures the
consistency of the Cluster resources.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>