Ensure the names in the tests factory are unique
See issue #411

The factories used in the unit tests have their name generated randomly.

Previously, because of this, the same name could be given to different
resources during one unit test. This could result in unexpected
conflicts during the test, which should not happen in production, as two
resources in the same namespace cannot have the same name.

This has been fixed, by changing the name generation algorithm. Instead
of taking a new name completely randomly and having the risk of taking
the same name several times, the generation algorithm ensures that each
name is unique and different from the ones used previously.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>