Simplify e2e tests by using convenience methods
In this commit we start using the convenience method
utils.create_default_environment() in favor of the methods
utils.create_multiple_cluster_environment() and
utils.create_simple_environment() methods, which are more
complicated to use.

We also use the convenience methods defined on
ResourceDefinition objects, such as app.update_resource()
and cluster.get_labels() for example. This avoids the more
complicated direct calls to the utils.run() method.

Signed-off-by: Karin Fetzer <karin.fetzer@cloudandheat.com>