Remove OpenVPN parts to prepare change to wireguard
See issue #384

The OpenVPN utility was leveraged to be able to connect through the
gateway to the Krake VMs and to the Minikube clusters deployed. However
the wireguard VPN is more simple to deploy and use, and thus will
replace OpenVPN in the Krake deployment.

This commit removes all OpenVPN-related elements, to have a blank page
to start with wireguard.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>