Create rok configuration suitable for wireguard
See issue #384

As wireguard is configured on the gateway to allow communications with
the Krake infrastructure, it means that the Krake API can be accessed
locally.

Thus, a rok configuration is generated locally, which points to the
Krake endpoint present on the private network accessible only through
wireguard. As TLS can be enabled, the needed certificates and key are
copied as well.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>