Add Kafka MetricProvider to use with KSQL
See issue #377

To allow fetching from a KSQL database fed by a Kafka Broker, a
MetricProvider for Kafka and its corresponding Provider in the Scheduler
have been added.

The Provider uses the KSQL pull query mechanism, see:
https://docs.confluent.io/platform/current/ksqldb/pull-queries-in-ccloud.html

It means that the Provider fetches the current value of a metric from a
KSQL table. More specifically, the value is one of the elements of a row
of this table. It is assumed that the chosen row is the one with the
chosen metric name as one of its attributes.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>