Support disabled TLS with docker-compose
When TLS was disabled on the Ansible host files, the docker-compose file
generation created an invalid file. Indeed, the file contains path to
the certificates generated for the TLS support. These certificates are
then mounted into the different components's container. If TLS is
disabled, they do not exist, thus the docker-compose utility cannot
bring the Krake component's up, and the utility fails.

The docker-compose file generation has now been modified to add only the
certificates when TLS is enabled. With this condition added, the
docker-compose utility can now start the Krake infrastructure with TLS
enabled or disabled.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>