Separate the Krake deployment with the test setup
See issue #371

Previously, when deploying Krake, some elements only used by the
integration tests were also inserted into the deployment infrastructure.
For instance some resources were bootstrapped in the database, or the
prometheus containers were started to have a prometheus instance
providing values for the Metrics in the tests.

As those elements were superfluous, they are now only added when
integration tests are started. This allows having a cleaner deployment
of Krake with the Ansible playbooks.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>