Ensure the kubernetes API session is closed
Previously, the kubernetes-asynciolibrary simply closed the session when
the object was deleted. So the reference to the session object was
simply set to None in the KubernetesClient utility when closing the
communication with the cluster.

With the latest updates of the library, the method for closing changed,
and a "close()" method is now available, while the session is not
closed anymore by the destructor. Thus, the previous method for closing
the session had to be changed. The KubernetesClient utility now calls
the "close()" method to ensure that the connection with the Kubernetes
cluster is terminated.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>