Add parameter for CORS Access-Control-Allow-Origin
CORS was enabled on Krake but the fields were set to be quite
unrestrictive. Mostly, the "Access-Control-Allow-Origin" was left to its
default value of '*'. With this setup, sending request through a browser
could be dangerous. A user could first connect to a valid website with
some allowed authentication token and send requests to Krake. Then the
user goes on a malicious website, which may be able to reuse the token.

To prevent this situation, the value for the
"Access-Control-Allow-Origin" field can now be set in the Krake API
configuration.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>