Fix Keystone unit tests remove and PySAML2 workaround
See issue #416

With the update of the pip package PySAML2, the Keystone instance took
a little more time to start. The difference was fine when running with
pytest, but exceeded the timeout when running the tests with tox.

Thus, the timeout was slightly increased, to allow the tests started by
tox to pass as well.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>