Add custom formatter class for argparse's help
When using `--help` to show the usage of the Krake components, we could
get the expected types of the parameters (str, int..) but not their
default value. This was thanks to the ``MetavarTypeHelpFormatter``
of the argparse module. Another potential provider is
``ArgumentDefaultsHelpFormatter``, which can display the default value
of the arguments, but not their type.

By using a custom formatter class which inherits both these provided
formatters, we can get the default value AND the expected type for each
argument.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>