Increase Kafka fixtures timeout for pipelines' tox
In the pipelines, when starting unit tests, tox is used instead of a
simple call to "pytest". Also, it is run in Docker container started by
the GitLab runners. These two differences with a simple "pytest" run
locally make the unit tests slower, and the timeouts used when
developing locally may not be high enough.

Thus, the timeouts have been increased. As these unit tests are usually
only run when the "--runslow" argument is set, this will not change much
when developing locally.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>