Fix timeouts in Kafka-related fixtures
See issue #419

In each one of the Kafka-related fixtures ("zookeeper", "kafka" and
"ksql"), a test is performed to verify that the instance has been
actually started and is reachable. In some cases, these verification led
to a timeout.

To verify if the first two services are ready ("zookeeper" and "kafka"),
a message is sent through sockets to the Zookeeper instance. The call
for reading the answer sometimes hanged.

By adding the "loop" argument, and changing the amount of bytes read
from the buffer to a clear limitation instead of waiting for EOF, the
timeout could be avoided.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>