Add workaround for make rsync faster in pipelines
See issue #418

Some pipelines were failing because of timeouts. It appeared that the
rsync task took sometimes more than an hour to finish, causing the rest
of the pipeline to be too late and exceeding the time limit.

To reduce the duration of the synchronize task (ansible task which
leverages rsync), the copy of the whole git repository is instead
separated into one synchronization task for each directory and files of
the root directory. The result is a drastic change in duration, from 1
hour to 30 seconds.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>