Refactor match methods of Scheduler
See issue #314

The static method match_cluster_constraints of the Scheduler was never
used anywhere inside the code. It was replaced earlier by another
function in `controller/scheduler/constraints.py`, but still present in
the Scheduler.

This left the static method match_project_constraints in the Scheduler,
while its pendant that was actually run for the Kubernetes clusters was
in the aforementionned file.

Thus, the unused method has been removed from the Scheduler, and the
matching function for the OS Project was moved to the constraints
module, to keep the code together.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>