Refactor the "rank" term into "score" in  scheduler
In the scheduler workflow, the resources (say "A") on which a new
resource "B" will be scheduled are first filtered, then ranked.

This second step is made by first: computing the score of each cluster,
then rank them and take the first one.

However, in the scheduler documentation and variables, in many places,
the term rank is used instead of score. This has been corrected.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>