Set e2e job to store components logs on failure
To allow easier debugging of the E2E tests job of the pipeline, the logs
of all Krake components are added as artifacts. However, due to the
`when` attribute which we do not set in the gitlab CI file, the default
value is used. It means that the artifacts (the logs in our case) are
only stored and made available on the GitLab portal when the job passed.

As the logs are only useful for debugging, the job is now set to store
the logs as artifacts only when the job fails.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>