Refactor test_reflector_retry to prevent randomness
See issue #354

The test test_reflector_retry which verified the behavior of the
Reflector's `__call__` method to retry the connection with a longer
delay each time was not deterministic. It sometimes failed because of
floating point computation, and assertions in the test for values.

To prevent this, the test is now only comparing the delays, and not
checking its value directly, to prevent edge cases with floating points.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>