Handle the Core resources in the garbage collector
See issue #313

The resources of the `core` API where previously not handled by the
Garbage Collector. The goal was to prevent "orphaned" resources, which
had references to deleted `Metric` or `MetricsProvider` resources for
instance.

In the end, it was decided to handle it anyway. Also this makes sense,
as misspelled metrics could also be added, and would need to be removed.
There is however a need for notifying the administrator when a metric is
being faulty. This will be added in another commit.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>