Add metrics issues in Cluster and Project status
See issue #429

Previously, if an error occurred during the fetching of the value of a
metric (metric or its provider not found in the database, provider
endpoint unreachable...), a simple log entry was added, but the Krake
user had no way of knowing what happened during the scheduling process,
so could not make decisions regarding the validity of the metrics.

Thus, the status of the `openstack.Projects` and the
`kubernetes.Cluster` now contains a mapping of the failing metrics to
the reason for the failure. This can allow a Krake user to get the
status of a Cluster or a Project, and to act accordingly.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>