Support label validation for invalid types
See issue #401

Previously, the labels of each resource were only tested against a
regular expression, as it was assumed the labels keys and values were
necessary strings. However, as we deserialize from JSON, other types
could be sent (boolean, integer, dictionaries...). The validation
function did not support other types than strings, and the validation
failed with a different error than `ValidationError` from `marshmallow`.
This would lead the API for instance to send a "500 Internal Error".

The `TypeError` kind of exception is now supported as well. This allows
the API to consider label keys or values with types other than strings
as invalid, without breaking the API with an exception and make it send
a `500` HTTP error.

Also, the validation of labels would stop as soon as one was considered
invalid (mostly when it did not match the regular expression). This
prevented the user to know all failing labels. This is now addressed:
any error for invalid key or value of the labels is shown to the user
when he/she attempted to send a malformed JSON to the API.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>