Invalidate Applications with empty constraints
See issue #397

As defined in the `data` module, `Application` instances could be
created with the `Constraints` field as `None`. Sending a `null` value
for this field in a JSON request the API would thus be considered valid.
The Application would afterwards be handled by other Controllers and
fail as the constraints parts of the resource would be missing.

Thus, the `spec.constraints` field of the Application is now set to be
mandatory, and it is not possible to leave it empty. The same principle
is applied to `constraints.cluster`.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>