Refactor Environment test utils context manager
The `get_resource_definition` class method of `Environment` was
changed to a normal method. Indeed, the method used no attributes from
the class, but needed one from the instances.

However, with this change, access to the `Environment` instance in the
test was necessary. Previously, when using an `Environment` instance
from the rak test utils as context manager, the output was the resources
managed. So no access to the instance of `Environment` and its methods
were possible.

The context manager of the `Environment` has thus been changed to
return the instance, which allows methods to be called on it from inside
the tests.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>