Prevent Controllers from failing after API error
See issue #344

When the Reflectors were fetching values from the API using the "list
and watch" mechanism, an error on the API would propagate to the
Reflector. It means that if the Reflectors were to receive an HTTP
`500`, the Reflectors would see it as an exception, raise it and make
their associated Controller fail. So a bug on the API could lead to the
whole Krake infrastructure breaking down.

The errors coming from the responses of the Krake API are now correctly
caught and logged. This is done on the Reflector, and allows it to carry
on its work, thus also prevents the Controller from failing.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>