Add support for resource lists in ApiObject repr
The `__repr__` method of the `ApiObject` class previously did not
support the instances of `<resource>List` classes. Trying to print such
an object would result in an `AttributeError` because these objects do
not have a `name` field in their metadata.

Support for the list has been added in the `__repr__` to prevent
exceptions and allow easier debugging of the Krake client library.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>