Allow deletion of Magnum Cluster during creation
See issue #330

When a Magnum Cluster was created, a request was sent to the Magnum API
by the worker, then it waited for the end of the creation process on the
OpenStack infrastructure. If, during this wait, a request to delete the
``MagnumCluster`` resource was sent on the Krake API, the actual
deletion could only be handled by a worker after the first worker
finished its job (success, failure, or deletion of the actual cluster).

To solve this, the original idea was to make use of an Observer on the
Magnum Controller, one for each Magnum cluster.

As a workaround, the current commit adds, inside the loop that
checks the current state of the actual cluster, a check for the deletion
state of the cluster on the Krake API. If it is in a to-be-deleted
state, the loop is exited and another worker can delete the cluster.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>