Add handling for unreachable providers for projects
The scheduler was updated to handle the case of some or all metrics
providers pointed by a Kubernetes cluster being unreachable. In that
case, the scheduler should still select a target for the Application.
This was fixed in 32483975a98325b20662c181ab943237080cd7ec.

The fix was however not applied to the OpenStack resources. The current
commit adds this specific case to the scheduler, so for the scheduling
of Magnum clusters on projects.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>