Add various fixes for the openstack API and rok
This commit contains modifications related to the Krake `openstack` API,
such as fixes on the Magnum Controller, on rok or on the `data` module
for resources of the `openstack` API. Those were necessary, but were
mostly updated to make the end-to-end test pass.

As the Magnum client sometimes returns a response that contains a list
of clusters instead of a cluster, a workaround was added to handle both
cases.

The constraints now have been modified for the `openstack` resources
just like for the commit 1a45356ae4676b7e50d7343b76c95bf3b80e2c0f did
with the `kubernetes` API resources.

The support of labels for the `openstack` resources in rok has also been
improved.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>