Add no-black option on the parser
See issue #303

When testing new templates, syntax errors can be added unintentionally.
Because black was used by default, the generation failed, as black
cannot handle syntax errors. So the generator simply printed an error
and stopped.

To prevent this, an option has been added to deactivate black when
necessary. This allows thus for easier debugging of the API generator.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>