Support blocking operation for app deletion
See issue #273

A new wait variable got added to the rok command for the kubernetes app deletion.
wait can be used on its own, which uses a default value or one out of a specific set of values can be picked.
This triggers the Krake API to only send a response, if the resource is in the state given with wait.
Internally, this is handled with a blocking decorator.
This decorator calls a database watcher to check the state of the to-be create / updated / deleted object.

Signed-off-by: Hannes Baum <hannes.baum@cloudandheat.com>