Create End-to-end tests for the creation and deletion of resources
In order to cover more code parts with the e2e tests, new tests for Role and RoleBinding are created.
Since these have not rok commands yet, we execute the tests with the help of the requests API, which is also under the hood of rok.
The test sequence is similar to the already existing tests in test_core.py.

Signed-off-by: Hannes Baum <hannes.baum@cloudandheat.com>