Add k8s secret as a storage for Complete hook confidentail information
Krake Complete hook stored confidentail information like token in
a Kubernetes ConfigMap resource. If a user had the permissions to get or
list Applications, he/she could obtain the token needed to delete an
Application mangled by the Complete hook.
This commit moves the Complete hook confidentail information to
the Kubernetes Secrets. It caused that the confidentail information
are not exposed in Application in a plain text.
It is a "kubernetes way" how to handle secret information.

Issue: #409

Signed-off-by: Matej Feder <matej.feder@dnation.cloud>