Add new shutdown hook
The new shutdown hook got added, which mangles the application on creation, if a `--hook shutdown` is used in rok for example.
Functionality is similar to the complete hook; new certificates are created and environment variables are given to the kubernetes cluster.
These are then bound with the application to deliver the needed information.

Signed-off-by: Hannes Baum <hannes.baum@cloudandheat.com>