Add RFC7807 Problem representation for failures on the HTTP layers
The failure reason on the API HTTP layer is stored as RFC 7807 Problem.
It is a way how to define a machine-readable and uniform details
of errors in an HTTP response.
See https://tools.ietf.org/html/rfc7807 for details.

Middleware factory :func:`problem_response` for HTTP exceptions was introduced.
Factory catches HttpProblemError or HTTPException based exception
transforming the exception text to the :class:`.helpers.HttpProblem`
(RFC 7807 Problem representation of failure) and reraising the exception.

Issue #455

Signed-off-by: Matej Feder <matej.feder@dnation.cloud>