Remove default Kubernetes namespace definition from observer schema
Most Kubernetes resources (e.g. pods, services, replication controllers,
and others) are in some namespaces. However, namespace resources are not
themselves in a namespace. And low-level resources, such as nodes and
persistentVolumes, are not in any namespace.
Therefore this commit removes the default Kubernetes namespace definition
from default as well as custom Observer Schema.

The issue may occur when the end-user wrongly defines the namespace
field in a non-namespaced resource or in its custom observer schema.
To prevent this, Krake skips the diff-calculation of namespace
field for non-namespaced resources.

The default Kubernetes namespace is applied only when the namespaced
Kubernetes resource does not have such a definition in its manifest.

Signed-off-by: Matej Feder <matej.feder@dnation.cloud>