Split the Kubernetes Controller
This commit splits the Kubernetes Controller in two separate controllers.
The Kubernetes Application Controller now explicitly manages Applications
which were scheduled on Kubernetes Clusters whereas the Kubernetes Cluster
Controller manages Kubernetes Clusters. The code base behind the Kubernetes
Controller with its tests and the documentation are effected by this. There
also are some changes to the file structure and naming of files.

Signed-off-by: Patrick Thiem <patrick.thiem@cloudandheat.com>