Fix application deletion stuck if deleted when cluster is disconnected
This commit fixes the following:
- :func:`application.handle_resource`did not
handle the :class:`ClientConnectorError` exception. This exception
is raised when the cluster API is unreachable.
This caused Krake could not delete the application from
the cluster and the application controller failed unexpectedly.
Hence, the :class:`ClientConnectorError` exception handling was added.
- The k8s application observer did not handle the :class:`ClientConnectorError`
exception as well. This caused Krake observer failed.
Hence, the :class:`ClientConnectorError` exception handling was added to
the observer logic as well. If the cluster is  unreachable
(:class:`ClientConnectorError`) the last known application status is
returned from observer's :func:`poll_resource`.
The application status should not change as we do not know the real
current state.

Issue #472

Signed-off-by: Matej Feder <matej.feder@dnation.cloud>