Update wireguard docs and MagnumCluster notification
See issue #492

The setup.rst file was updated regarding the wireguard documentation.
A guide was included on how to set up wireguard and which specific
wg-key should be included in the wg_<peer_name>.conf.
The MagnumCluster is depreciated. As it is not actively developed
the documentation was adjusted. All suggested changes were applied.
A notification for the user was added.
Nothing was deleted. If the MagnumCluster will be used in the future,
remove the added notification.