Add options to setup  logging to krake config generation
When using the krake_generate_config script to generate configuraitons,
it is now possible to set the logging level and the handler to use for
logging.

This commit allows it to set the logging level, as well of the logging
handler, via the group_vars in Ansible.

It is also possible to set these individually for each krake componant
itself.

Signed-off-by: Toni Finger <toni.finger@cloudandheat.com>