Fix minikube related paths
1. Home dir reference
The setup guide in the README intructs to retrieve the minikube config, certs
and keys from one's user home but references it wrongly. Instead of
`/home/USER/.minikube/...` `~/.minikube/...` is more accurate. Note that
`/home/$USER/.minikube/...` would also be wrong because a users home dir below
/home/ only by convention.

2. Cert/key file paths in copied minikube config
The minikube CA cert, client key and cert are copied but their paths were not
adjusted in the config.

Co-authored-by: @felix.mue