Schedule to cluster if scores are equal to cloud
During scheduling, if two equal maxima are found and they're cloud and cluster, it was previously randomly decided which one would taken for scheduling. This is obviously bad, if a cluster was previously created on a cloud and they're both the same. Therefore, the cluster is now preferred during scheduling of equal valued objects.

Signed-off-by: Hannes Baum <hannes.baum@cloudandheat.com>