Replace implicit `null` values with deletionState
The `deleted` property used `null` to indicate, that a resource is not
yet scheduled for deletion. This is problematic for several reasons:

- Using `null` to represent anything but "this property is null" is
not recommended. Everybody using this property essentially has to guess,
what `null` means.

- In some programming languages / serialization frameworks also
missing properties are deserialized to null.

This commit replaces it with `deletionState` containing the boolean
`deleted` and a `deletionTime` (datetime) if `deleted` is `true`.

Signed-off-by: Felix Müller <felix.mueller@cloudandheat.com>

. Zeilen,